5.15@home 23.24

+++ src/main/java/com/psylife/entity/DimensionEntity.java
@@ -19,29 +19,29 @@ public class DimensionEntity implements Serializable {
 	private int choice;
 	private String answer;
 	//
-	private int loyalty;
+	private double loyalty;
 	//
-	private int positive;
+	private double positive;
 	//
-	private int responsibility;
+	private double responsibility;
 	//
-	private int morality;
+	private double morality;
 	//
-	private int thinking;
+	private double thinking;
 	//
-	private int plan;
+	private double plan;
 	//
-	private int innovation;
+	private double innovation;
 	//
-	private int teamwork;
+	private double teamwork;
 	//
-	private int communication;
+	private double communication;
 	//
-	private int strain;
+	private double strain;
 	//
-	private int details;
+	private double details;
 	//
-	private int potential;
+	private double potential;
 
 	/**
 	 * @return the id
@@ -83,7 +83,7 @@ public class DimensionEntity implements Serializable {
 	/**
 	 * @return the loyalty
 	 */
-	public int getLoyalty() {
+	public double getLoyalty() {
 		return loyalty;
 	}
 
@@ -91,14 +91,14 @@ public class DimensionEntity implements Serializable {
 	 * @param loyalty
 	 *            the loyalty to set
 	 */
-	public void setLoyalty(int loyalty) {
+	public void setLoyalty(double loyalty) {
 		this.loyalty = loyalty;
 	}
 
 	/**
 	 * @return the positive
 	 */
-	public int getPositive() {
+	public double getPositive() {
 		return positive;
 	}
 
@@ -106,14 +106,14 @@ public class DimensionEntity implements Serializable {
 	 * @param positive
 	 *            the positive to set
 	 */
-	public void setPositive(int positive) {
+	public void setPositive(double positive) {
 		this.positive = positive;
 	}
 
 	/**
 	 * @return the responsibility
 	 */
-	public int getResponsibility() {
+	public double getResponsibility() {
 		return responsibility;
 	}
 
@@ -121,14 +121,14 @@ public class DimensionEntity implements Serializable {
 	 * @param responsibility
 	 *            the responsibility to set
 	 */
-	public void setResponsibility(int responsibility) {
+	public void setResponsibility(double responsibility) {
 		this.responsibility = responsibility;
 	}
 
 	/**
 	 * @return the morality
 	 */
-	public int getMorality() {
+	public double getMorality() {
 		return morality;
 	}
 
@@ -136,14 +136,14 @@ public class DimensionEntity implements Serializable {
 	 * @param morality
 	 *            the morality to set
 	 */
-	public void setMorality(int morality) {
+	public void setMorality(double morality) {
 		this.morality = morality;
 	}
 
 	/**
 	 * @return the thinking
 	 */
-	public int getThinking() {
+	public double getThinking() {
 		return thinking;
 	}
 
@@ -151,14 +151,14 @@ public class DimensionEntity implements Serializable {
 	 * @param thinking
 	 *            the thinking to set
 	 */
-	public void setThinking(int thinking) {
+	public void setThinking(double thinking) {
 		this.thinking = thinking;
 	}
 
 	/**
 	 * @return the plan
 	 */
-	public int getPlan() {
+	public double getPlan() {
 		return plan;
 	}
 
@@ -166,14 +166,14 @@ public class DimensionEntity implements Serializable {
 	 * @param plan
 	 *            the plan to set
 	 */
-	public void setPlan(int plan) {
+	public void setPlan(double plan) {
 		this.plan = plan;
 	}
 
 	/**
 	 * @return the innovation
 	 */
-	public int getInnovation() {
+	public double getInnovation() {
 		return innovation;
 	}
 
@@ -181,14 +181,14 @@ public class DimensionEntity implements Serializable {
 	 * @param innovation
 	 *            the innovation to set
 	 */
-	public void setInnovation(int innovation) {
+	public void setInnovation(double innovation) {
 		this.innovation = innovation;
 	}
 
 	/**
 	 * @return the teamwork
 	 */
-	public int getTeamwork() {
+	public double getTeamwork() {
 		return teamwork;
 	}
 
@@ -196,14 +196,14 @@ public class DimensionEntity implements Serializable {
 	 * @param teamwork
 	 *            the teamwork to set
 	 */
-	public void setTeamwork(int teamwork) {
+	public void setTeamwork(double teamwork) {
 		this.teamwork = teamwork;
 	}
 
 	/**
 	 * @return the communication
 	 */
-	public int getCommunication() {
+	public double getCommunication() {
 		return communication;
 	}
 
@@ -211,14 +211,14 @@ public class DimensionEntity implements Serializable {
 	 * @param communication
 	 *            the communication to set
 	 */
-	public void setCommunication(int communication) {
+	public void setCommunication(double communication) {
 		this.communication = communication;
 	}
 
 	/**
 	 * @return the strain
 	 */
-	public int getStrain() {
+	public double getStrain() {
 		return strain;
 	}
 
@@ -226,14 +226,14 @@ public class DimensionEntity implements Serializable {
 	 * @param strain
 	 *            the strain to set
 	 */
-	public void setStrain(int strain) {
+	public void setStrain(double strain) {
 		this.strain = strain;
 	}
 
 	/**
 	 * @return the details
 	 */
-	public int getDetails() {
+	public double getDetails() {
 		return details;
 	}
 
@@ -241,14 +241,14 @@ public class DimensionEntity implements Serializable {
 	 * @param details
 	 *            the details to set
 	 */
-	public void setDetails(int details) {
+	public void setDetails(double details) {
 		this.details = details;
 	}
 
 	/**
 	 * @return the potential
 	 */
-	public int getPotential() {
+	public double getPotential() {
 		return potential;
 	}
 
@@ -256,7 +256,7 @@ public class DimensionEntity implements Serializable {
 	 * @param potential
 	 *            the potential to set
 	 */
-	public void setPotential(int potential) {
+	public void setPotential(double potential) {
 		this.potential = potential;
 	}
